+++ braintree/webhook_testing_gateway.py
@@ -71,6 +71,8 @@ class WebhookTestingGateway(object):
             return self.__dispute_won_sample_xml(id)
         elif kind == WebhookNotification.Kind.DisputeAccepted:
             return self.__dispute_accepted_sample_xml(id)
+        elif kind == WebhookNotification.Kind.DisputeAutoAccepted:
+            return self.__dispute_auto_accepted_sample_xml(id)
         elif kind == WebhookNotification.Kind.DisputeDisputed:
             return self.__dispute_disputed_sample_xml(id)
         elif kind == WebhookNotification.Kind.DisputeExpired:
@@ -244,6 +246,12 @@ class WebhookTestingGateway(object):
         else:
             return self.__new_dispute_accepted_sample_xml(id)
 
+    def __dispute_auto_accepted_sample_xml(self, id):
+        if id == "legacy_dispute_id":
+            return self.__old_dispute_auto_accepted_sample_xml(id)
+        else:
+            return self.__new_dispute_auto_accepted_sample_xml(id)
+
     def __dispute_disputed_sample_xml(self, id):
         if id == "legacy_dispute_id":
             return self.__old_dispute_disputed_sample_xml(id)
@@ -323,6 +331,25 @@ class WebhookTestingGateway(object):
               <reply-by-date type="date">2014-03-21</reply-by-date>
               <kind>chargeback</kind>
               <status>accepted</status>
+              <reason>fraud</reason>
+              <id>%s</id>
+              <transaction>
+                <id>%s</id>
+                <amount>250.00</amount>
+              </transaction>
+              <date-opened type="date">2014-03-28</date-opened>
+            </dispute>
+        """ % (id, id)
+
+    def __old_dispute_auto_accepted_sample_xml(self, id):
+        return """
+            <dispute>
+              <amount>250.00</amount>
+              <currency-iso-code>USD</currency-iso-code>
+              <received-date type="date">2014-03-01</received-date>
+              <reply-by-date type="date">2014-03-21</reply-by-date>
+              <kind>chargeback</kind>
+              <status>auto_accepted</status>
               <reason>fraud</reason>
               <id>%s</id>
               <transaction>
@@ -562,6 +589,51 @@ class WebhookTestingGateway(object):
             </status-history>
             <status-history>
               <status>accepted</status>
+              <timestamp type="datetime">2017-06-16T20:44:41Z</timestamp>
+            </status-history>
+          </status-history>
+          <evidence type="array"/>
+          <transaction>
+            <id>%s</id>
+            <amount>100.00</amount>
+            <created-at>2017-06-21T20:44:41Z</created-at>
+            <order-id nil="true"/>
+            <purchase-order-number nil="true"/>
+            <payment-instrument-subtype>Visa</payment-instrument-subtype>
+          </transaction>
+          <date-opened type=\"date\">2014-03-28</date-opened>
+        </dispute>
+        """ % (id, id)
+
+    def __new_dispute_auto_accepted_sample_xml(self, id):
+        return """
+        <dispute>
+          <id>%s</id>
+          <amount>100.00</amount>
+          <amount-disputed>100.00</amount-disputed>
+          <amount-won>95.00</amount-won>
+          <case-number>CASE-12345</case-number>
+          <created-at type="datetime">2017-06-16T20:44:41Z</created-at>
+          <currency-iso-code>USD</currency-iso-code>
+          <forwarded-comments nil="true"/>
+          <kind>chargeback</kind>
+          <merchant-account-id>ytnlulaloidoqwvzxjrdqputg</merchant-account-id>
+          <reason>fraud</reason>
+          <reason-code nil="true"/>
+          <reason-description nil="true"/>
+          <received-date type="date">2016-02-15</received-date>
+          <reference-number>REF-9876</reference-number>
+          <reply-by-date type="date">2016-02-22</reply-by-date>
+          <status>auto_accepted</status>
+          <updated-at type="datetime">2017-06-16T20:44:41Z</updated-at>
+          <original-dispute-id>9qde5qgp</original-dispute-id>
+          <status-history type="array">
+            <status-history>
+              <status>open</status>
+              <timestamp type="datetime">2017-06-15T20:44:41Z</timestamp>
+            </status-history>
+            <status-history>
+              <status>auto_accepted</status>
               <timestamp type="datetime">2017-06-16T20:44:41Z</timestamp>
             </status-history>
           </status-history>