+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Remove mock dev dependency (Thanks @arthurzam)
 * Add `ExcessiveRetry` to `GatewayRejectionReason`
 * Add `pre_dispute_program` to `Dispute` and `DisputeSearch`
+* Add `AutoAccepted` status to `Dispute`
+* Add `DisputeAutoAccepted` to `WebhookNotification.Kind`
 
 ## 4.17.1
 * Prepare http request before setting url to resolve issue where dot segments get normalized